refactor(BlackJack) : State 인터페이스에서 Open Class로 변경
