+++ src/main/kotlin/blackjack/domain/participants/user/Guest.kt
@@ -1,7 +1,7 @@
 package blackjack.domain.participants.user
 
+import blackjack.domain.state.FirstTurn
 import blackjack.domain.state.State
-import blackjack.domain.state.inTurn.FirstTurn
 
 class Guest(
     val bettingMoney: Money = Money(10),
