+++ src/main/kotlin/blackjack/domain/state/Hit.kt
@@ -1,13 +1,9 @@
-package blackjack.domain.state.inTurn
+package blackjack.domain.state
 
 import blackjack.domain.card.Card
 import blackjack.domain.card.Cards
-import blackjack.domain.state.Outcome
-import blackjack.domain.state.State
-import blackjack.domain.state.endTurn.Bust
-import blackjack.domain.state.endTurn.Stay
 
-class Hit(cards: Cards) : InTurn(cards) {
+class Hit(cards: Cards) : State(cards) {
     override fun draw(card: Card): State {
         val newCards: Cards = cards + card
         return when {
