+++ src/test/kotlin/blackjack/domain/participants/user/DealerTest.kt
@@ -7,8 +7,8 @@ import blackjack.domain.state.Fixtures.CLOVER_QUEEN
 import blackjack.domain.state.Fixtures.CLOVER_SEVEN
 import blackjack.domain.state.Fixtures.CLOVER_SIX
 import blackjack.domain.state.Fixtures.CLOVER_TEN
-import blackjack.domain.state.endTurn.Stay
-import blackjack.domain.state.inTurn.Hit
+import blackjack.domain.state.Hit
+import blackjack.domain.state.Stay
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
