+++ src/test/kotlin/blackjack/domain/state/BustTest.kt
@@ -1,4 +1,4 @@
-package blackjack.domain.state.endTurn
+package blackjack.domain.state
 
 import blackjack.domain.card.Cards
 import blackjack.domain.state.Fixtures.CLOVER_ACE
@@ -6,7 +6,6 @@ import blackjack.domain.state.Fixtures.CLOVER_JACK
 import blackjack.domain.state.Fixtures.CLOVER_KING
 import blackjack.domain.state.Fixtures.CLOVER_QUEEN
 import blackjack.domain.state.Fixtures.CLOVER_TWO
-import blackjack.domain.state.State
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
